Good start on keeping session state alive; removing as many references to global app
Works to keep session state alive, doesn't fully work to prompt users to maintain session/logout/log back in.

I've been relying on the only-in-debug-mode global reference to the app; if isDebug is set to false, that's not available. Start at removing as many of these as possible, replacing with topic publish/subscribe, refactoring where appropriate.

Refactored the way layers are created, so that we don't re-create them every time if user adds then removes a layer.