Work-around for rasters
We now have map services with raster layers, which work as services, but not when attempting to load as individual RasterLayers. This is a start on what should work to support raster layers, but results in Error: Output format not supported. Code: 400 errors. As a temporary work-around, the Add to Map button in layer browser is disabled for these layers, and in search results a button to load the map service is now shown.

Also added code to handle a layer loading, but not yet loaded, to prevent user from clicking again to load a layer that is taking too long to load. (So far this has only happened when attempting, and failing, to load raster layers, but it doesn't hurt to have this "loadPending" capability.)