Address auto-complete
At a crossroads: the jQueryUI autocomplete has almost the exact behavior I want, but renders poorly in an otherwise Dojo-centered UI. I'm either going to switch it out to use the native Esri search widget embedded in the form, or start hacking jQuery UI css to solve the rendering problems. Committing this revision as a placeholder to return to if the first option doesn't pan out.