More tweaks for #45
I noticed that if the first thing the user does is click the search button, nothing shows up. (The validator function does run, but it doesn't tirgger display.) Added code to set focus to the search textbox on startup, which (hackily) forces it to display if user then clicks the button or types enter.