Follow-up after merging changes for issue #52 into publicSiteMap branch
The publicSiteMap branch was created from estMap a while ago, and the Search-and-print-issue-52 branch just recently, so changes made to estMap since the last time anything was merged over get swept in. This is a partial revert, including switching back to externally hosted FA (can't use the secure EST hosted version on public site map) and meaningless changes to package-lock.json.