#67 loading project tree
Adding support for menus in dynamic folders, to add all layers on/off toggle at the folder level. This is useful beyond just the need for it in EST project service layers, and could be used with any dynamic map service that incorporates folders. Note also the notes added about ignoreDynamicGroupVisibility parameter.