#71 Adding more validation to feature editor
One of the ways to trigger analysis creation failures is to have invalid attributes or geometry on a feature--0 buffer on a point or line, missing name (perhaps, but would cause other problems as well), or null shape. The validation added here focuses on the name (required, not duplicated) and buffer distance (0 min or null for poly, (min of 1). Included are commented out earlier attempts that would undo any change a user makes to feature attributes, the way the existing prod version handles a user entering a duplicate name. The current approach is to let bad data happen, but alert the user, and block moving onto the next step in the wizard until problems are solved. It still needs to check for null shapes, but those, I think, can only be generated via extract, so I'll probably handle under #72.  This also includes a comment on where I left off trying to trace other possible causes of analysis area creation failure.