Issue 67 Work In Progress
Committing work in progress before simplifying it. Extra complexity is included in this version to handle cases where one alt is in draft and another is not in draft. This case shouldn't happen, and team's consensus is to correct the data for the few (older) projects to keep the alt statuses synced (at least with regard to < 6 vs >=6 statuses. Next version will assume all alts are in the same status.