Issue #67 nearing completion
Fully functional ability to provide option to load draft when available and allowed. Added loading overlay for step where it's checking MapDAO--it's reasonably fast enough on the server to not warrant a loading overlay, but if I'm demoing it locally there's a long pause. Now that it's been added to the mixin we may want to use it on other steps.

Determined that the correct layer visibility for projects services doesn't need to include 0, with the current settings of the folder checkbox visibility. In fact, if it does include 0, first draw of a dynamically loaded project service shows the termini, and some mysterious label, that goes away after user toggles any layer's visibility. You can see the layer ids change from [0, 6, 7, 8] to something like [5, 6, 7, 8] or [7, 8]--whatever the user changes the 0 isn't part of the array.